core-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
psycopg[binary,pool]==3.2.3
alembic==1.13.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pyotp==2.9.0
httpx==0.27.2
